fix: removed room related code
