+++ lib/pages/chat_page.dart
@@ -4,7 +4,6 @@ import 'package:my_chat_app/components/user_avatar.dart';
 import 'package:my_chat_app/cubits/messages/messages_cubit.dart';
 import 'package:my_chat_app/models/message.dart';
 import 'package:my_chat_app/utils/constants.dart';
-import 'package:my_chat_app/utils/messages_provider.dart';
 import 'package:timeago/timeago.dart';
 
 /// Page to chat with someone.
@@ -13,12 +12,10 @@ import 'package:timeago/timeago.dart';
 class ChatPage extends StatelessWidget {
   const ChatPage({Key? key}) : super(key: key);
 
-  static Route<void> route(String roomId) {
-    final messagesProvider = MessagesProvider();
+  static Route<void> route() {
     return MaterialPageRoute(
       builder: (context) => BlocProvider<MessagesCubit>(
-        create: (context) => MessagesCubit(messagesProvider: messagesProvider)
-          ..setMessagesListener(roomId),
+        create: (context) => MessagesCubit()..setMessagesListener(),
         child: const ChatPage(),
       ),
     );
